Fixing memory leaks in Foundation.

+++ Frameworks/Foundation/NSDictionary.mm
@@ -630,7 +630,7 @@ static int _NSDict_SortedKeysHelper(id key1, id key2, void* context) {
 /**
  @Status Interoperable
 */
-- (uint64_t)fileSize {
+- (unsigned long long)fileSize {
     __int64 ret = [[self objectForKey:NSFileSize] intValue];
 
     return ret;
